Add parseFromData helper function
Add parseFormData helper function to provide a convenient way to access
uploaded form data in an application using airship. The helper function
uses the parseRequestBody function from the wai-extra package. This
parseRequestBody function provides the ability to select different
backends for file uploads, but currently parseFormData does not expose
this and just uses the in-memory backend.