Set location header prior to post processing
Set location header prior to calling the post processing function for
PostProcessRedirect handling to give the post processing function the
option to override the Location header based on the logic in the post
processing function. Also add a replaceResponseHeader helper to provide
a convenient way to overwrite the value of a header that may already
exist in the ReponseHeaders list.