Provide configuration API for Airship apps.
This emulates Warp's approach towards web server settings in that it
provides a data type (`AirshipConfig`) with a hidden constructor, which
is modified through a default value and a set of functions to read/write
individual settings. We piggyback on the microlens library to provide these functions.

Please note that this is a breaking change, as it adds a parameter to
`resourceToWai`.

Fixes #49.