Implement custom 404 responses
Airship provides a means to specify a resource used for requests for
undefined routes. This allows for specifying a custom response body to
accompany 404 status codes. Airship does not actually do anything with
these at the present time. This change updates the handling of the L7
node in the decision tree to use the contentTypesProvided by the 404
resource to return a custom response body as intended by the user.