Remove space leaks in the Airship-Trace header.
This patch contains three optimizations:

The three optimizations are as follows:
1) We stop using RWST, as its Writer component leaks space. We replace
it with an RST monad, extracted from the Snap framework, that provides
assured strictness and an inlined >>= implementation. In addition, we
use modify' rather than modify to ensure that the change is applied
strictly (this yielded a small but significant increase in throughput).

2) We use (:) to add a new item onto the internal list of traces,
rather than constructing a new list out of the item and using mappend.
Though this constructs the list of traces in reverse order, we reverse
said list before sending it to the client.

3) We use ByteString values in the internal trace list so as to avoid
the overhead of an unnecessary decodeUtf8 (since the header value is
a ByteString anyway).