Setup Hackage releases from tags.
Travis CI supports uploading the results of a build to Hackage as a new version. When you do a new release via the github project eg https://github.com/helium/airship/releases it will do a build and attempt to push the version indetified in the cabal file to hackage.